Split out new renderer, add tests
